async_device: Reorganize certification tests
